Art and Bugfixes
Fixed tile art so it has shadows on the correct side. Added a pointer graphic that lines up with the goal tile when tiles are at rest. Increased the number of max Lear tiles on screen at once. Fixed a bug where the price display check would cause an overflow if there were no upgrades of that type remaining.